add more funcs to staging_helpers
