+++ src/staging/staging_main.py
@@ -1,6 +1,5 @@
 """The main file for the staging and validation module."""
 import logging
-from numpy import random
 from typing import Callable, Tuple
 from datetime import datetime
 import pandas as pd
@@ -10,7 +9,7 @@ from src.staging import spp_parser, history_loader
 from src.staging import spp_snapshot_processing as processing
 from src.staging import validation as val
 from src.staging import pg_conversion as pg
-from src.staging.staging_helpers import postcode_topup
+from src.staging.staging_helpers import postcode_topup, update_ref_list, fix_anon_data
 from src.utils.wrappers import time_logger_wrap
 
 StagingMainLogger = logging.getLogger(__name__)
@@ -89,37 +88,6 @@ def load_snapshot_feather(feather_file, read_feather):
     snapdata = read_feather(feather_file)
     StagingMainLogger.info(f"{feather_file} loaded")
     return snapdata
-
-
-def fix_anon_data(responses_df, config):
-    """
-    Fixes anonymised snapshot data for use in the DevTest environment.
-
-    This function adds an "instance" column to the provided DataFrame, and populates
-    it with zeros. It also adds a "selectiontype" column with random values of "P",
-    "C", or "L", and a "cellnumber" column with random values from the "seltype_list"
-    in the configuration.
-
-    This fix is necessary because the anonymised snapshot data currently used in the
-    DevTest environment does not include the "instance" column. This fix should be
-    removed when new anonymised data is provided.
-
-    Args:
-        responses_df (pandas.DataFrame): The DataFrame containing the anonymised
-        snapshot data.
-        config (dict): A dictionary containing configuration details.
-
-    Returns:
-        pandas.DataFrame: The fixed DataFrame with the added "instance",
-        "selectiontype", and "cellnumber" columns.
-    """
-    responses_df["instance"] = 0
-    col_size = responses_df.shape[0]
-    random.seed(seed=42)
-    responses_df["selectiontype"] = random.choice(["P", "C", "L"], size=col_size)
-    cellno_list = config["devtest"]["seltype_list"]
-    responses_df["cellnumber"] = random.choice(cellno_list, size=col_size)
-    return responses_df
 
 
 def load_val_snapshot_json(snapshot_path, load_json, config, network_or_hdfs):
@@ -572,7 +540,7 @@ def run_staging(
         val.validate_data_with_schema(ref_list_817, schema_path)
         StagingMainLogger.info("reference list mapper File Loaded Successfully...")
         # update longform references that should be on the reference list
-        full_responses = val.update_ref_list(full_responses, ref_list_817)
+        full_responses = update_ref_list(full_responses, ref_list_817)
     else:
         StagingMainLogger.info("Skipping loding the reference list mapper File.")
