+++ src/staging/validation.py
@@ -631,48 +631,3 @@ def validate_cora_df(df: pd.DataFrame) -> pd.DataFrame:
 
     except ValueError as ve:
         raise ValueError("cora status mapper validation failed: " + str(ve))
-
-
-@exception_wrap
-def update_ref_list(full_df: pd.DataFrame, ref_list_df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Update long form references that should be on the reference list.
-
-    For the first year (processing 2022 data) only, several references
-    should have been designated on the "reference list", ie, should have been
-    assigned cellnumber = 817, but were wrongly assigned a different cellnumber.
-
-    Args:
-        full_df (pd.DataFrame): The full_responses dataframe
-        ref_list_df (pd.DataFrame): The mapper containing updates for the cellnumber
-    Returns:
-        df (pd.DataFrame): with cellnumber and selectiontype cols updated.
-    """
-    ref_list_filtered = ref_list_df.loc[
-        (ref_list_df.formtype == 1) & (ref_list_df.cellnumber != 817)
-    ]
-    df = pd.merge(
-        full_df,
-        ref_list_filtered[["reference", "cellnumber"]],
-        how="outer",
-        on="reference",
-        suffixes=("", "_new"),
-        indicator=True,
-    )
-    # check no items in the reference list mapper are missing from the full responses
-    missing_refs = df.loc[df["_merge"] == "right_only"]
-    if not missing_refs.empty:
-        msg = (
-            "The following references in the reference list mapper are not in the data:"
-        )
-        raise ValueError(msg + str(missing_refs.reference.unique()))
-
-    # update cellnumber and selectiontype where there is a match
-    match_cond = df["_merge"] == "both"
-    df = df.copy()
-    df.loc[match_cond, "cellnumber"] = 817
-    df.loc[match_cond, "selectiontype"] = "L"
-
-    df = df.drop(["_merge", "cellnumber_new"], axis=1)
-
-    return df
